+++ src/infrastructure/transport/grpc/index.ts
@@ -2,3 +2,4 @@ export { createSdkChannel } from './sdk-channel';
 export { createSdkClient } from './sdk-client';
 export { createSdkMetadata } from './sdk-metadata';
 export { createSdkMiddleware } from './sdk-middleware';
+export { UnaryLimitResolver } from './unary-limit-resolver';